fix(app): start the repair of the fsk app
Over the past 2 years, there has been a number of automatic bumps of
versions as recommended by bots either checking for CVEs or
recommending updates to the app. In order to this, new Python versions
have released, packages such Flask itself and flask-jwt-extended has
also changed.

This has led to this app breaking and not functioning anymore. This
commit is an attempt to fix this and continue with active development of
the application